Refactor: DTO DAO Layer 수정 [박한솔]
