+++ basilium-server/src/main/java/basilium/basiliumserver/BasiliumServerApplication.java
@@ -45,4 +45,5 @@ public class BasiliumServerApplication {
 			}
 		};
 	}
+
 }
