+++ basilium-server/src/main/java/basilium/basiliumserver/controller/user/NormalUserController.java
@@ -1,11 +1,12 @@
 package basilium.basiliumserver.controller.user;
 
 import basilium.basiliumserver.auth.support.AuthUser;
-import basilium.basiliumserver.domain.payment.OrderPaymentRequest;
+import basilium.basiliumserver.domain.deliveryInfo.DeliveryInfo;
 import basilium.basiliumserver.domain.user.*;
-import basilium.basiliumserver.service.Like.LikeService;
-import basilium.basiliumserver.service.payment.PaymentService;
-import basilium.basiliumserver.service.shoppingCart.ShoppingCartService;
+import basilium.basiliumserver.service.DTO.user.AccessTokenResponse;
+import basilium.basiliumserver.service.DTO.user.NormalUserInfoDTO;
+import basilium.basiliumserver.service.DTO.user.RefreshTokenRequest;
+import basilium.basiliumserver.service.DTO.user.UserModifiedInfo;
 import basilium.basiliumserver.service.user.NormalUserService;
 import io.micrometer.common.util.StringUtils;
 import lombok.RequiredArgsConstructor;
