+++ basilium-server/src/main/java/basilium/basiliumserver/error/exception/BaseException.java
@@ -1,4 +1,4 @@
-package basilium.basiliumserver.common.exception;
+package basilium.basiliumserver.error.exception;
 
 import lombok.Getter;
