+++ basilium-server/src/main/java/basilium/basiliumserver/error/exception/Status.java
@@ -1,4 +1,4 @@
-package basilium.basiliumserver.common.exception;
+package basilium.basiliumserver.error.exception;
 
 public enum Status {
