+++ basilium-server/src/main/java/basilium/basiliumserver/service/DTO/payment/kafkaPaymentInventory/ProductUpdateListener.java
@@ -1,4 +1,4 @@
-package basilium.basiliumserver.domain.payment.kafkaPaymentInventory;
+package basilium.basiliumserver.service.DTO.payment.kafkaPaymentInventory;
 
 import basilium.basiliumserver.configuration.kafkaMQ.KafkaConfig;
 import basilium.basiliumserver.service.product.ProductService;
