+++ basilium-server/src/main/java/basilium/basiliumserver/service/DTO/user/UserModifiedInfo.java
@@ -1,4 +1,4 @@
-package basilium.basiliumserver.domain.user;
+package basilium.basiliumserver.service.DTO.user;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
