+++ basilium-server/src/main/java/basilium/basiliumserver/service/like/LikeService.java
@@ -1,4 +1,4 @@
-package basilium.basiliumserver.service.Like;
+package basilium.basiliumserver.service.like;
 
 
 import basilium.basiliumserver.domain.product.Product;
