+++ basilium-server/src/main/resources/application.properties
@@ -6,6 +6,17 @@ spring.datasource.url=jdbc:mysql://localhost:3306/basilium?serverTimezone=UTC&ch
 spring.datasource.username=basilium
 spring.datasource.password=pjhcsol
 
+# JPA ??
+# ddl-auto (create,update) ??
+spring.jpa.show-sql=true
+spring.jpa.hibernate.ddl-auto=create
+#spring.jpa.hibernate.ddl-auto=create-drop
+
+# schema.sql and data.sql into dataSource ??
+#spring.sql.init.schema-locations=classpath:schema.sql
+#spring.sql.init.data-locations=classpath:data.sql
+#spring.sql.init.mode=always
+
 server.port = 8080
 #????? ??? ??? ??? ??
 #spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQLDialect
@@ -44,10 +55,6 @@ spring.datasource.hikari.maximum-pool-size=24
 
 #thymeleaf
 spring.thymeleaf.check-template-location=false
-
-# JPA ??
-spring.jpa.show-sql=true
-spring.jpa.hibernate.ddl-auto=create
 
 
 # userImage upload dir
